STY: moved OMNI HRO local funcs
Moved the OMNI HRO local functions to a separate instrument module.  Also fixed a bug in the clean routine, which now allows for array-like inputs.
